Renamed BaleenException::invalidObjectException and created tests for it.
