+++ src/Exception/BaleenException.php
@@ -26,14 +26,14 @@ namespace Baleen\Migrations\Exception;
 class BaleenException extends \Exception
 {
     /**
-     * throwInvalidObjectException
+     * invalidObjectException
      *
      * @param $object
      * @param $expected
      *
      * @throws static
      */
-    public static function throwInvalidObjectException($object, $expected)
+    public static function invalidObjectException($object, $expected)
     {
         throw new static(sprintf(
             'Expected an object of type "%s", but got "%s" instead.',
